Other iterations/video-analytics-system/local_script/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.3
ultralytics==8.1.11
insightface==0.7.3
onnxruntime==1.16.3
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.2.0

# For BoT-SORT tracking
filterpy==1.4.5
scipy==1.11.4
lap==0.4.0
